Renaming Files and adding Js doc comments
